refactor(lexer): reorganize tests into focused test modules
- Created src/lexer/tests/ directory with 7 test modules:
  * basic_tests.rs - basic tokenization, words, operators (30 tests)
  * quote_tests.rs - quote handling (7 tests)
  * expansion_tests.rs - parameter, arithmetic, command substitution (81 tests)
  * redirection_tests.rs - redirection tokenization (48 tests)
  * tilde_tests.rs - tilde expansion (18 tests)
  * alias_tests.rs - alias expansion (4 tests)
  * edge_case_tests.rs - placeholder for future edge cases

- Reduced src/lexer/mod.rs from 3,531 lines to 1,348 lines
- All 151 lexer tests pass
- Full test suite passes (669 tests)

The lexer implementation is now at a manageable size (~1,348 lines)
and doesn't require further extraction.
