+++ src/lexer/mod.rs
@@ -1346,2187 +1346,4 @@ pub fn expand_aliases(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::Mutex;
-
-    // Mutex to serialize tests that modify environment variables
-    static ENV_LOCK: Mutex<()> = Mutex::new(());
-
-    /// Helper function to expand tokens like the executor does
-    /// This simulates what happens at execution time
-    fn expand_tokens(tokens: Vec<Token>, shell_state: &mut ShellState) -> Vec<Token> {
-        let mut result = Vec::new();
-        for token in tokens {
-            match token {
-                Token::Word(word) => {
-                    // Use the executor's expansion logic
-                    let expanded = crate::executor::expand_variables_in_string(&word, shell_state);
-                    // If expansion results in empty string, and it was a command substitution that produced no output,
-                    // we might need to skip adding it (for test_command_substitution_empty_output)
-                    if !expanded.is_empty() || !word.starts_with("$(") {
-                        result.push(Token::Word(expanded));
-                    }
-                }
-                other => result.push(other),
-            }
-        }
-        result
-    }
-
-    #[test]
-    fn test_basic_word() {
-        let shell_state = ShellState::new();
-        let result = lex("ls", &shell_state).unwrap();
-        assert_eq!(result, vec![Token::Word("ls".to_string())]);
-    }
-
-    #[test]
-    fn test_multiple_words() {
-        let shell_state = ShellState::new();
-        let result = lex("ls -la", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("ls".to_string()),
-                Token::Word("-la".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_pipe() {
-        let shell_state = ShellState::new();
-        let result = lex("ls | grep txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("ls".to_string()),
-                Token::Pipe,
-                Token::Word("grep".to_string()),
-                Token::Word("txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_redirections() {
-        let shell_state = ShellState::new();
-        let result = lex("printf hello > output.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("printf".to_string()),
-                Token::Word("hello".to_string()),
-                Token::RedirOut,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_append_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("printf hello >> output.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("printf".to_string()),
-                Token::Word("hello".to_string()),
-                Token::RedirAppend,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_input_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("cat < input.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::RedirIn,
-                Token::Word("input.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_double_quotes() {
-        let shell_state = ShellState::new();
-        let result = lex("echo \"hello world\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_single_quotes() {
-        let shell_state = ShellState::new();
-        let result = lex("echo 'hello world'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_variable_expansion() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "expanded_value".to_string());
-        let tokens = lex("echo $TEST_VAR", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("expanded_value".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_variable_expansion_nonexistent() {
-        let shell_state = ShellState::new();
-        let result = lex("echo $TEST_VAR2", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("$TEST_VAR2".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_empty_variable() {
-        let shell_state = ShellState::new();
-        let result = lex("echo $", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("$".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_mixed_quotes_and_variables() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("USER", "alice".to_string());
-        let tokens = lex("echo \"Hello $USER\"", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("Hello alice".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_unclosed_double_quote() {
-        // Lexer doesn't handle unclosed quotes as errors, just treats as literal
-        let shell_state = ShellState::new();
-        let result = lex("echo \"hello", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_empty_input() {
-        let shell_state = ShellState::new();
-        let result = lex("", &shell_state).unwrap();
-        assert_eq!(result, Vec::<Token>::new());
-    }
-
-    #[test]
-    fn test_only_spaces() {
-        let shell_state = ShellState::new();
-        let result = lex("   ", &shell_state).unwrap();
-        assert_eq!(result, Vec::<Token>::new());
-    }
-
-    #[test]
-    fn test_complex_pipeline() {
-        let shell_state = ShellState::new();
-        let result = lex(
-            "cat input.txt | grep \"search term\" > output.txt",
-            &shell_state,
-        )
-        .unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::Word("input.txt".to_string()),
-                Token::Pipe,
-                Token::Word("grep".to_string()),
-                Token::Word("search term".to_string()),
-                Token::RedirOut,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_if_tokens() {
-        let shell_state = ShellState::new();
-        let result = lex("if true; then printf yes; fi", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::If,
-                Token::Word("true".to_string()),
-                Token::Semicolon,
-                Token::Then,
-                Token::Word("printf".to_string()),
-                Token::Word("yes".to_string()),
-                Token::Semicolon,
-                Token::Fi,
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_dollar_paren() {
-        let shell_state = ShellState::new();
-        let result = lex("echo $(pwd)", &shell_state).unwrap();
-        // The output will vary based on current directory, but should be a single Word token
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        assert!(matches!(result[1], Token::Word(_)));
-    }
-
-    #[test]
-    fn test_command_substitution_backticks() {
-        let shell_state = ShellState::new();
-        let result = lex("echo `pwd`", &shell_state).unwrap();
-        // The output will vary based on current directory, but should be a single Word token
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        assert!(matches!(result[1], Token::Word(_)));
-    }
-
-    #[test]
-    fn test_command_substitution_with_arguments() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $(echo hello world)", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_backticks_with_arguments() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo `echo hello world`", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_failure_fallback() {
-        let shell_state = ShellState::new();
-        let result = lex("echo $(nonexistent_command)", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("$(nonexistent_command)".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_backticks_failure_fallback() {
-        let shell_state = ShellState::new();
-        let result = lex("echo `nonexistent_command`", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("`nonexistent_command`".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_with_variables() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "test_value".to_string());
-        let tokens = lex("echo $(echo $TEST_VAR)", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test_value".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_in_assignment() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("MY_VAR=$(echo hello)", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        // The lexer treats MY_VAR= as a single word, then appends the substitution result
-        assert_eq!(result, vec![Token::Word("MY_VAR=hello".to_string())]);
-    }
-
-    #[test]
-    fn test_command_substitution_backticks_in_assignment() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("MY_VAR=`echo hello`", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        // The lexer correctly separates MY_VAR= from the substitution result
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("MY_VAR=".to_string()),
-                Token::Word("hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_with_quotes() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo \"$(echo hello world)\"", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_backticks_with_quotes() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo \"`echo hello world`\"", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_empty_output() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $(true)", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        // true produces no output, so we get just "echo"
-        assert_eq!(result, vec![Token::Word("echo".to_string())]);
-    }
-
-    #[test]
-    fn test_command_substitution_multiple_spaces() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $(echo 'hello   world')", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello   world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_with_newlines() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $(printf 'hello\nworld')", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello\nworld".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_command_substitution_special_characters() {
-        let shell_state = ShellState::new();
-        let result = lex("echo $(echo '$#@^&*()')", &shell_state).unwrap();
-        println!("Special chars test result: {:?}", result);
-        // The actual output shows $#@^&*() but test expects $#@^&*()
-        // This might be due to shell interpretation of # as comment
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        assert!(matches!(result[1], Token::Word(_)));
-    }
-
-    #[test]
-    fn test_nested_command_substitution() {
-        // Note: Current implementation doesn't support nested substitution
-        // This test documents the current behavior
-        let shell_state = ShellState::new();
-        let result = lex("echo $(echo $(pwd))", &shell_state).unwrap();
-        // The inner $(pwd) is not processed because it's part of the command string
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        assert!(matches!(result[1], Token::Word(_)));
-    }
-
-    #[test]
-    fn test_command_substitution_in_pipeline() {
-        let shell_state = ShellState::new();
-        let result = lex("$(echo hello) | cat", &shell_state).unwrap();
-        println!("Pipeline test result: {:?}", result);
-        assert_eq!(result.len(), 3);
-        assert!(matches!(result[0], Token::Word(_)));
-        assert_eq!(result[1], Token::Pipe);
-        assert_eq!(result[2], Token::Word("cat".to_string()));
-    }
-
-    #[test]
-    fn test_command_substitution_with_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("$(echo hello) > output.txt", &shell_state).unwrap();
-        assert_eq!(result.len(), 3);
-        assert!(matches!(result[0], Token::Word(_)));
-        assert_eq!(result[1], Token::RedirOut);
-        assert_eq!(result[2], Token::Word("output.txt".to_string()));
-    }
-
-    #[test]
-    fn test_variable_in_quotes_with_pipe() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("PATH", "/usr/bin:/bin".to_string());
-        let tokens = lex("echo \"$PATH\" | tr ':' '\\n'", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("/usr/bin:/bin".to_string()),
-                Token::Pipe,
-                Token::Word("tr".to_string()),
-                Token::Word(":".to_string()),
-                Token::Word("\\n".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_expand_aliases_simple() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_alias("ll", "ls -l".to_string());
-        let tokens = vec![Token::Word("ll".to_string())];
-        let result = expand_aliases(tokens, &shell_state, &mut HashSet::new()).unwrap();
-        assert_eq!(
-            result,
-            vec![Token::Word("ls".to_string()), Token::Word("-l".to_string())]
-        );
-    }
-
-    #[test]
-    fn test_expand_aliases_with_args() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_alias("ll", "ls -l".to_string());
-        let tokens = vec![
-            Token::Word("ll".to_string()),
-            Token::Word("/tmp".to_string()),
-        ];
-        let result = expand_aliases(tokens, &shell_state, &mut HashSet::new()).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("ls".to_string()),
-                Token::Word("-l".to_string()),
-                Token::Word("/tmp".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_expand_aliases_no_alias() {
-        let shell_state = ShellState::new();
-        let tokens = vec![Token::Word("ls".to_string())];
-        let result = expand_aliases(tokens.clone(), &shell_state, &mut HashSet::new()).unwrap();
-        assert_eq!(result, tokens);
-    }
-
-    #[test]
-    fn test_expand_aliases_chained() {
-        // Test that chained aliases work correctly: a -> b -> a (command)
-        // This is NOT recursion in bash - it expands a to b, then b to a (the command),
-        // and then tries to execute command 'a' which doesn't exist.
-        let mut shell_state = ShellState::new();
-        shell_state.set_alias("a", "b".to_string());
-        shell_state.set_alias("b", "a".to_string());
-        let tokens = vec![Token::Word("a".to_string())];
-        let result = expand_aliases(tokens, &shell_state, &mut HashSet::new());
-        // Should succeed and expand to just "a" (the command, not the alias)
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), vec![Token::Word("a".to_string())]);
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_simple() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $((2 + 3))", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("5".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_with_variables() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("x", "10".to_string());
-        shell_state.set_var("y", "20".to_string());
-        let tokens = lex("echo $((x + y * 2))", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("50".to_string()) // 10 + 20 * 2 = 50
-            ]
-        );
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_comparison() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $((5 > 3))", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("1".to_string()) // true
-            ]
-        );
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_complex() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("a", "3".to_string());
-        let tokens = lex("echo $((a * 2 + 5))", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("11".to_string()) // 3 * 2 + 5 = 11
-            ]
-        );
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_unmatched_parentheses() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $((2 + 3", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        // The unmatched parentheses should remain as literal, possibly with formatting
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        // Accept either the original or a formatted version with the literal kept
-        let second_token = &result[1];
-        if let Token::Word(s) = second_token {
-            assert!(
-                s.starts_with("$((") && s.contains("2") && s.contains("3"),
-                "Expected unmatched arithmetic to be kept as literal, got: {}",
-                s
-            );
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_arithmetic_expansion_division_by_zero() {
-        let mut shell_state = ShellState::new();
-        let tokens = lex("echo $((5 / 0))", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        // Division by zero produces an error message
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        // The second token should contain an error message about division by zero
-        if let Token::Word(s) = &result[1] {
-            assert!(
-                s.contains("Division by zero"),
-                "Expected division by zero error, got: {}",
-                s
-            );
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_parameter_expansion_simple() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello world".to_string());
-        let result = lex("echo ${TEST_VAR}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_unset_variable() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ${UNSET_VAR}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![Token::Word("echo".to_string()), Token::Word("".to_string())]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_default() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ${UNSET_VAR:-default}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("default".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_default_set_variable() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "value".to_string());
-        let result = lex("echo ${TEST_VAR:-default}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("value".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_assign_default() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ${UNSET_VAR:=default}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("default".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_alternative() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "value".to_string());
-        let result = lex("echo ${TEST_VAR:+replacement}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("replacement".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_alternative_unset() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ${UNSET_VAR:+replacement}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![Token::Word("echo".to_string()), Token::Word("".to_string())]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_substring() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello world".to_string());
-        let result = lex("echo ${TEST_VAR:6}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_substring_with_length() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello world".to_string());
-        let result = lex("echo ${TEST_VAR:0:5}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_length() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello".to_string());
-        let result = lex("echo ${#TEST_VAR}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("5".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_remove_shortest_prefix() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "prefix_hello".to_string());
-        let result = lex("echo ${TEST_VAR#prefix_}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_remove_longest_prefix() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "prefix_prefix_hello".to_string());
-        let result = lex("echo ${TEST_VAR##prefix_}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("prefix_hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_remove_shortest_suffix() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello_suffix".to_string());
-        let result = lex("echo ${TEST_VAR%suffix}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello_".to_string()) // Fixed: should be "hello_" not "hello"
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_remove_longest_suffix() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello_suffix_suffix".to_string());
-        let result = lex("echo ${TEST_VAR%%suffix}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello_suffix_".to_string()) // Fixed: correct result is "hello_suffix_"
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_substitute() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello world".to_string());
-        let result = lex("echo ${TEST_VAR/world/universe}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello universe".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_substitute_all() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello world world".to_string());
-        let result = lex("echo ${TEST_VAR//world/universe}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello universe universe".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_mixed_with_regular_variables() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("VAR1", "value1".to_string());
-        shell_state.set_var("VAR2", "value2".to_string());
-        let tokens = lex("echo $VAR1 and ${VAR2}", &shell_state).unwrap();
-        let result = expand_tokens(tokens, &mut shell_state);
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("value1".to_string()),
-                Token::Word("and".to_string()),
-                Token::Word("value2".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_in_double_quotes() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("TEST_VAR", "hello".to_string());
-        let result = lex("echo \"Value: ${TEST_VAR}\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("Value: hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_parameter_expansion_error_unset() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ${UNSET_VAR:?error message}", &shell_state);
-        // Should fall back to literal syntax on error
-        assert!(result.is_ok());
-        let tokens = result.unwrap();
-        assert_eq!(tokens.len(), 3);
-        assert_eq!(tokens[0], Token::Word("echo".to_string()));
-        assert_eq!(tokens[1], Token::Word("${UNSET_VAR:?error}".to_string()));
-        assert_eq!(tokens[2], Token::Word("message}".to_string()));
-    }
-
-    #[test]
-    fn test_parameter_expansion_complex_expression() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("PATH", "/usr/bin:/bin:/usr/local/bin".to_string());
-        let result = lex("echo ${PATH#/usr/bin:}", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("/bin:/usr/local/bin".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_local_keyword() {
-        let shell_state = ShellState::new();
-        let result = lex("local myvar", &shell_state).unwrap();
-        assert_eq!(result, vec![Token::Local, Token::Word("myvar".to_string())]);
-    }
-
-    #[test]
-    fn test_local_keyword_in_function() {
-        let shell_state = ShellState::new();
-        let result = lex("local var=value", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![Token::Local, Token::Word("var=value".to_string())]
-        );
-    }
-
-    #[test]
-    fn test_single_quotes_with_semicolons() {
-        // Test that semicolons inside single quotes are preserved as part of the string
-        let shell_state = ShellState::new();
-        let result = lex("trap 'echo \"A\"; echo \"B\"' EXIT", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("trap".to_string()),
-                Token::Word("echo \"A\"; echo \"B\"".to_string()),
-                Token::Word("EXIT".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_double_quotes_with_semicolons() {
-        // Test that semicolons inside double quotes are preserved as part of the string
-        let shell_state = ShellState::new();
-        let result = lex("echo \"command1; command2\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("command1; command2".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_semicolons_outside_quotes() {
-        // Test that semicolons outside quotes still work as command separators
-        let shell_state = ShellState::new();
-        let result = lex("echo hello; echo world", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::Semicolon,
-                Token::Word("echo".to_string()),
-                Token::Word("world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_here_document_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("cat << EOF", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::RedirHereDoc("EOF".to_string(), false)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_here_string_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("cat <<< \"hello world\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::RedirHereString("hello world".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_here_document_with_quoted_delimiter() {
-        let shell_state = ShellState::new();
-        let result = lex("command << 'EOF'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirHereDoc("EOF".to_string(), true) // Quoted delimiter
-            ]
-        );
-    }
-
-    #[test]
-    fn test_here_string_without_quotes() {
-        let shell_state = ShellState::new();
-        let result = lex("grep <<< pattern", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("grep".to_string()),
-                Token::RedirHereString("pattern".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_redirections_mixed() {
-        let shell_state = ShellState::new();
-        let result = lex(
-            "cat < input.txt <<< \"fallback\" > output.txt",
-            &shell_state,
-        )
-        .unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::RedirIn,
-                Token::Word("input.txt".to_string()),
-                Token::RedirHereString("fallback".to_string()),
-                Token::RedirOut,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_unquoted() {
-        let _lock = ENV_LOCK.lock().unwrap();
-        let shell_state = ShellState::new();
-        let home = env::var("HOME").unwrap_or_else(|_| "/home/user".to_string());
-        let result = lex("echo ~", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![Token::Word("echo".to_string()), Token::Word(home)]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_single_quoted() {
-        let shell_state = ShellState::new();
-        let result = lex("echo '~'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_double_quoted() {
-        let shell_state = ShellState::new();
-        let result = lex("echo \"~\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_mixed_quotes() {
-        let _lock = ENV_LOCK.lock().unwrap();
-        let shell_state = ShellState::new();
-        let home = env::var("HOME").unwrap_or_else(|_| "/home/user".to_string());
-        let result = lex("echo ~ '~' \"~\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word(home),
-                Token::Word("~".to_string()),
-                Token::Word("~".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_pwd() {
-        let mut shell_state = ShellState::new();
-
-        // Set PWD variable
-        let test_pwd = "/test/current/dir";
-        shell_state.set_var("PWD", test_pwd.to_string());
-
-        let result = lex("echo ~+", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word(test_pwd.to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_oldpwd() {
-        let mut shell_state = ShellState::new();
-
-        // Set OLDPWD variable
-        let test_oldpwd = "/test/old/dir";
-        shell_state.set_var("OLDPWD", test_oldpwd.to_string());
-
-        let result = lex("echo ~-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word(test_oldpwd.to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_pwd_unset() {
-        let _lock = ENV_LOCK.lock().unwrap();
-        let shell_state = ShellState::new();
-
-        // When PWD is not set, ~+ should expand to current directory
-        let result = lex("echo ~+", &shell_state).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-
-        // The second token should be a valid path (either from env::current_dir or literal ~+)
-        if let Token::Word(path) = &result[1] {
-            // Should either be a path or the literal ~+
-            assert!(path.starts_with('/') || path == "~+");
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_tilde_expansion_oldpwd_unset() {
-        // Lock to prevent parallel tests from interfering with environment variables
-        let _lock = ENV_LOCK.lock().unwrap();
-
-        // Save and clear OLDPWD
-        let original_oldpwd = env::var("OLDPWD").ok();
-        unsafe {
-            env::remove_var("OLDPWD");
-        }
-
-        let shell_state = ShellState::new();
-
-        // When OLDPWD is not set, ~- should remain as literal
-        let result = lex("echo ~-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~-".to_string())
-            ]
-        );
-
-        // Restore OLDPWD
-        unsafe {
-            if let Some(oldpwd) = original_oldpwd {
-                env::set_var("OLDPWD", oldpwd);
-            }
-        }
-    }
-
-    #[test]
-    fn test_tilde_expansion_pwd_in_quotes() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("PWD", "/test/dir".to_string());
-
-        // Single quotes should prevent expansion
-        let result = lex("echo '~+'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~+".to_string())
-            ]
-        );
-
-        // Double quotes should also prevent expansion
-        let result = lex("echo \"~+\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~+".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_oldpwd_in_quotes() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("OLDPWD", "/test/old".to_string());
-
-        // Single quotes should prevent expansion
-        let result = lex("echo '~-'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~-".to_string())
-            ]
-        );
-
-        // Double quotes should also prevent expansion
-        let result = lex("echo \"~-\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~-".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_mixed() {
-        let _lock = ENV_LOCK.lock().unwrap();
-        let mut shell_state = ShellState::new();
-        let home = env::var("HOME").unwrap_or_else(|_| "/home/user".to_string());
-        shell_state.set_var("PWD", "/current".to_string());
-        shell_state.set_var("OLDPWD", "/previous".to_string());
-
-        let result = lex("echo ~ ~+ ~-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word(home),
-                Token::Word("/current".to_string()),
-                Token::Word("/previous".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_not_at_start() {
-        let mut shell_state = ShellState::new();
-        shell_state.set_var("PWD", "/test".to_string());
-
-        // Tilde should not expand when not at start of word
-        let result = lex("echo prefix~+", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("prefix~+".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_username() {
-        let shell_state = ShellState::new();
-
-        // Test with root username (special case: /root instead of /home/root)
-        let result = lex("echo ~root", &shell_state).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-
-        // The expansion should either be /root or literal ~root (if /root doesn't exist)
-        if let Token::Word(path) = &result[1] {
-            assert!(path == "/root" || path == "~root");
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_tilde_expansion_username_with_path() {
-        let shell_state = ShellState::new();
-
-        // Test ~username/path expansion
-        let result = lex("echo ~root/documents", &shell_state).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-
-        // Should expand to /root/documents or ~root/documents
-        if let Token::Word(path) = &result[1] {
-            assert!(path == "/root/documents" || path == "~root/documents");
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_tilde_expansion_nonexistent_user() {
-        let shell_state = ShellState::new();
-
-        // Test with a username that definitely doesn't exist
-        let result = lex("echo ~nonexistentuser12345", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~nonexistentuser12345".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_username_in_quotes() {
-        let shell_state = ShellState::new();
-
-        // Single quotes should prevent expansion
-        let result = lex("echo '~root'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~root".to_string())
-            ]
-        );
-
-        // Double quotes should also prevent expansion
-        let result = lex("echo \"~root\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("~root".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_tilde_expansion_mixed_with_username() {
-        let _lock = ENV_LOCK.lock().unwrap();
-        let mut shell_state = ShellState::new();
-        let home = env::var("HOME").unwrap_or_else(|_| "/home/user".to_string());
-        shell_state.set_var("PWD", "/current".to_string());
-
-        // Test mixing different tilde expansions
-        let result = lex("echo ~ ~+ ~root", &shell_state).unwrap();
-        assert_eq!(result.len(), 4);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-        assert_eq!(result[1], Token::Word(home));
-        assert_eq!(result[2], Token::Word("/current".to_string()));
-
-        // The ~root expansion depends on whether /root exists
-        if let Token::Word(path) = &result[3] {
-            assert!(path == "/root" || path == "~root");
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    #[test]
-    fn test_tilde_expansion_username_with_special_chars() {
-        let shell_state = ShellState::new();
-
-        // Test that special characters terminate username collection
-        let result = lex("echo ~user@host", &shell_state).unwrap();
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0], Token::Word("echo".to_string()));
-
-        // Should try to expand ~user and then append @host
-        if let Token::Word(path) = &result[1] {
-            // The path should contain @host at the end
-            assert!(path.contains("@host") || path == "~user@host");
-        } else {
-            panic!("Expected Word token");
-        }
-    }
-
-    // ===== File Descriptor Redirection Tests =====
-
-    #[test]
-    fn test_fd_output_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>errors.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOut(2, "errors.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_input_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<input.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdIn(3, "input.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_append_redirection() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>>errors.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdAppend(2, "errors.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_duplication_output() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>&1", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdDup(2, 1)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_duplication_input() {
-        let shell_state = ShellState::new();
-        let result = lex("command 0<&3", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdDup(0, 3)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_close_output() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>&-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdClose(2)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_close_input() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<&-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdClose(3)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_read_write() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<>file.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdInOut(3, "file.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_read_write_default() {
-        let shell_state = ShellState::new();
-        let result = lex("command <>file.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdInOut(0, "file.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_multiple_fd_redirections() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>err.log 3<input.txt 4>>append.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOut(2, "err.log".to_string()),
-                Token::RedirectFdIn(3, "input.txt".to_string()),
-                Token::RedirectFdAppend(4, "append.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_redirection_with_pipe() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>&1 | grep error", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdDup(2, 1),
-                Token::Pipe,
-                Token::Word("grep".to_string()),
-                Token::Word("error".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_numbers_0_through_9() {
-        let shell_state = ShellState::new();
-
-        // Test fd 0
-        let result = lex("cmd 0<file", &shell_state).unwrap();
-        assert_eq!(result[1], Token::RedirectFdIn(0, "file".to_string()));
-
-        // Test fd 9
-        let result = lex("cmd 9>file", &shell_state).unwrap();
-        assert_eq!(result[1], Token::RedirectFdOut(9, "file".to_string()));
-    }
-
-    #[test]
-    fn test_fd_swap_pattern() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3>&1 1>&2 2>&3 3>&-", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdDup(3, 1),
-                Token::RedirectFdDup(1, 2),
-                Token::RedirectFdDup(2, 3),
-                Token::RedirectFdClose(3)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_backward_compat_simple_output() {
-        let shell_state = ShellState::new();
-        let result = lex("echo hello > output.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::RedirOut,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_backward_compat_simple_input() {
-        let shell_state = ShellState::new();
-        let result = lex("cat < input.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("cat".to_string()),
-                Token::RedirIn,
-                Token::Word("input.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_backward_compat_append() {
-        let shell_state = ShellState::new();
-        let result = lex("echo hello >> output.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::RedirAppend,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_with_spaces() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2> errors.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOut(2, "errors.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_no_space() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>errors.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOut(2, "errors.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_dup_to_self() {
-        let shell_state = ShellState::new();
-        let result = lex("command 1>&1", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdDup(1, 1)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_stderr_to_stdout() {
-        let shell_state = ShellState::new();
-        let result = lex("ls /nonexistent 2>&1", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("ls".to_string()),
-                Token::Word("/nonexistent".to_string()),
-                Token::RedirectFdDup(2, 1)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_stdout_to_stderr() {
-        let shell_state = ShellState::new();
-        let result = lex("echo error 1>&2", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("error".to_string()),
-                Token::RedirectFdDup(1, 2)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_combined_redirections() {
-        let shell_state = ShellState::new();
-        let result = lex("command >output.txt 2>&1", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirOut,
-                Token::Word("output.txt".to_string()),
-                Token::RedirectFdDup(2, 1)
-            ]
-        );
-    }
-
-    #[test]
-    fn test_fd_with_variable_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>$LOGFILE", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOut(2, "$LOGFILE".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_invalid_fd_dup_no_target() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>&", &shell_state);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("expected fd number or '-' after >&")
-        );
-    }
-
-    #[test]
-    fn test_invalid_fd_close_input_no_dash() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<&", &shell_state);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("expected fd number or '-' after <&")
-        );
-    }
-
-    #[test]
-    fn test_fd_inout_no_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<>", &shell_state);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("expected filename after <>"));
-    }
-
-    #[test]
-    fn test_fd_output_no_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>", &shell_state);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("expected filename after >"));
-    }
-
-    #[test]
-    fn test_fd_input_no_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3<", &shell_state);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("expected filename after <"));
-    }
-
-    #[test]
-    fn test_fd_append_no_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>>", &shell_state);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("expected filename after >>"));
-    }
-
-    // ===== Bang (!) Context-Aware Tests =====
-
-    #[test]
-    fn test_bang_as_argument() {
-        // ! should be treated as a regular word when not at command start
-        let shell_state = ShellState::new();
-        let result = lex("echo !", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_as_argument_middle() {
-        let shell_state = ShellState::new();
-        let result = lex("echo hello !", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_as_argument_multiple() {
-        let shell_state = ShellState::new();
-        let result = lex("echo ! ! !", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string()),
-                Token::Word("!".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_at_command_start() {
-        // ! at command start should be Token::Bang for negation
-        let shell_state = ShellState::new();
-        let result = lex("! echo hello", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Bang,
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_semicolon() {
-        let shell_state = ShellState::new();
-        let result = lex("echo hello; ! false", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::Semicolon,
-                Token::Bang,
-                Token::Word("false".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_newline() {
-        let shell_state = ShellState::new();
-        let result = lex("echo hello\n! false", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello".to_string()),
-                Token::Newline,
-                Token::Bang,
-                Token::Word("false".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_and_operator() {
-        let shell_state = ShellState::new();
-        let result = lex("true && ! false", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("true".to_string()),
-                Token::And,
-                Token::Bang,
-                Token::Word("false".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_or_operator() {
-        let shell_state = ShellState::new();
-        let result = lex("false || ! false", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("false".to_string()),
-                Token::Or,
-                Token::Bang,
-                Token::Word("false".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_pipe() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test | ! grep test", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::Pipe,
-                Token::Word("!".to_string()),
-                Token::Word("grep".to_string()),
-                Token::Word("test".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_in_single_quotes() {
-        let shell_state = ShellState::new();
-        let result = lex("echo '!'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_in_double_quotes() {
-        let shell_state = ShellState::new();
-        let result = lex("echo \"!\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_mixed_contexts() {
-        // Test both negation and argument usage in same line
-        let shell_state = ShellState::new();
-        let result = lex("! echo ! && echo !", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Bang,
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string()),
-                Token::And,
-                Token::Word("echo".to_string()),
-                Token::Word("!".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_after_then() {
-        let shell_state = ShellState::new();
-        let result = lex("if true; then ! false; fi", &shell_state).unwrap();
-        assert!(result.contains(&Token::Bang));
-        // Verify Bang comes after Then
-        let then_pos = result.iter().position(|t| t == &Token::Then).unwrap();
-        let bang_pos = result.iter().position(|t| t == &Token::Bang).unwrap();
-        assert!(bang_pos > then_pos);
-    }
-
-    #[test]
-    fn test_bang_after_else() {
-        let shell_state = ShellState::new();
-        let result = lex("if false; then echo a; else ! true; fi", &shell_state).unwrap();
-        assert!(result.contains(&Token::Bang));
-        // Verify Bang comes after Else
-        let else_pos = result.iter().position(|t| t == &Token::Else).unwrap();
-        let bang_pos = result.iter().position(|t| t == &Token::Bang).unwrap();
-        assert!(bang_pos > else_pos);
-    }
-
-    #[test]
-    fn test_bang_after_do() {
-        let shell_state = ShellState::new();
-        let result = lex("while true; do ! false; done", &shell_state).unwrap();
-        assert!(result.contains(&Token::Bang));
-        // Verify Bang comes after Do
-        let do_pos = result.iter().position(|t| t == &Token::Do).unwrap();
-        let bang_pos = result.iter().position(|t| t == &Token::Bang).unwrap();
-        assert!(bang_pos > do_pos);
-    }
-
-    #[test]
-    fn test_bang_in_subshell() {
-        let shell_state = ShellState::new();
-        let result = lex("(! echo test)", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::LeftParen,
-                Token::Bang,
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::RightParen
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_in_command_group() {
-        let shell_state = ShellState::new();
-        let result = lex("{ ! echo test; }", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::LeftBrace,
-                Token::Bang,
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::Semicolon,
-                Token::RightBrace
-            ]
-        );
-    }
-
-    #[test]
-    fn test_bang_as_part_of_word() {
-        // When ! is part of a word (no space before it), it should be included in the word
-        let shell_state = ShellState::new();
-        let result = lex("echo hello!", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("hello!".to_string())
-            ]
-        );
-    }
-
-    // ===== OutputClobber (>|) Tests =====
-
-    #[test]
-    fn test_output_clobber_basic() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test >| output.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::RedirOutClobber,
-                Token::Word("output.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_with_fd_number() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test 2>| errors.log", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::RedirectFdOutClobber(2, "errors.log".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_with_fd_number_no_space() {
-        let shell_state = ShellState::new();
-        let result = lex("command 3>|file.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOutClobber(3, "file.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_missing_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test >|", &shell_state);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("expected filename after >|"));
-    }
-
-    #[test]
-    fn test_output_clobber_with_quoted_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test >| \"output file.txt\"", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::RedirOutClobber,
-                Token::Word("output file.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_with_fd_and_quoted_filename() {
-        let shell_state = ShellState::new();
-        let result = lex("command 2>| 'error log.txt'", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirectFdOutClobber(2, "error log.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_multiple_fds() {
-        let shell_state = ShellState::new();
-        let result = lex("command >| out.txt 2>| err.txt", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("command".to_string()),
-                Token::RedirOutClobber,
-                Token::Word("out.txt".to_string()),
-                Token::RedirectFdOutClobber(2, "err.txt".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_with_pipe() {
-        let shell_state = ShellState::new();
-        let result = lex("echo test >| output.txt | cat", &shell_state).unwrap();
-        assert_eq!(
-            result,
-            vec![
-                Token::Word("echo".to_string()),
-                Token::Word("test".to_string()),
-                Token::RedirOutClobber,
-                Token::Word("output.txt".to_string()),
-                Token::Pipe,
-                Token::Word("cat".to_string())
-            ]
-        );
-    }
-
-    #[test]
-    fn test_output_clobber_fd_0_through_9() {
-        let shell_state = ShellState::new();
-        
-        // Test fd 0 (unusual but valid)
-        let result = lex("cmd 0>| file", &shell_state).unwrap();
-        assert_eq!(result[1], Token::RedirectFdOutClobber(0, "file".to_string()));
-        
-        // Test fd 9
-        let result = lex("cmd 9>| file", &shell_state).unwrap();
-        assert_eq!(result[1], Token::RedirectFdOutClobber(9, "file".to_string()));
-    }
-
-    #[test]
-    fn test_output_clobber_consistency_with_regular_redirect() {
-        let shell_state = ShellState::new();
-        
-        // Regular redirect without fd
-        let result1 = lex("echo test > output.txt", &shell_state).unwrap();
-        // Clobber redirect without fd
-        let result2 = lex("echo test >| output.txt", &shell_state).unwrap();
-        
-        // Both should have same structure, just different redirect token
-        assert_eq!(result1.len(), result2.len());
-        assert_eq!(result1[0], result2[0]); // echo
-        assert_eq!(result1[1], result2[1]); // test
-        assert_eq!(result1[3], result2[3]); // output.txt
-    }
-}
+mod tests;
