+++ src/lexer/tests/edge_case_tests.rs
@@ -1,0 +1,5 @@
+//! Edge case and error condition tests
+
+// This module is intentionally minimal as most edge cases are already
+// covered in the other test modules. This is a placeholder for future
+// edge case tests that don't fit neatly into other categories.
