+++ src/lexer/tests/mod.rs
@@ -1,0 +1,9 @@
+//! Tests for the lexer module
+
+mod basic_tests;
+mod quote_tests;
+mod expansion_tests;
+mod redirection_tests;
+mod tilde_tests;
+mod alias_tests;
+mod edge_case_tests;
